# kernel/internal/gateway/queryapi/server_gen.go
//go:build go1.22

// Package queryapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package queryapi

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"github.com/oapi-codegen/runtime"
)

const (
	ServiceTokenScopes  = "serviceToken.Scopes"
	UserAssertionScopes = "userAssertion.Scopes"
)

// Defines values for ErrorCode.
const (
	ErrorCodeCeilingExceeded        ErrorCode = "ceiling_exceeded"
	ErrorCodeConditionLimitExceeded ErrorCode = "condition_limit_exceeded"
	ErrorCodeNotFound               ErrorCode = "not_found"
	ErrorCodeNotGroupable           ErrorCode = "not_groupable"
	ErrorCodeNotOrderable           ErrorCode = "not_orderable"
	ErrorCodeOperatorNotAllowed     ErrorCode = "operator_not_allowed"
	ErrorCodeSchemaInvalid          ErrorCode = "schema_invalid"
	ErrorCodeScoreTypeMismatch      ErrorCode = "score_type_mismatch"
	ErrorCodeUnauthorized           ErrorCode = "unauthorized"
	ErrorCodeUnknownField           ErrorCode = "unknown_field"
)

// Error defines model for Error.
type Error struct {
	Code ErrorCode `json:"code"`

	// Detail Names the offending field/constant where applicable.
	Detail  *map[string]interface{} `json:"detail,omitempty"`
	Message string                  `json:"message"`
}

// ErrorCode defines model for Error.Code.
type ErrorCode string

// QueryResponse defines model for QueryResponse.
type QueryResponse struct {
	// Cursor Opaque keyset cursor; present iff a next page exists (row queries only).
	Cursor *string `json:"cursor,omitempty"`

	// Data Row queries return canonical entities (span/trace/score shapes, subject to field-level redaction); aggregation queries return one object per group (groupBy values + aggregation aliases).
	Data     []interface{}       `json:"data"`
	Stats    QueryResponse_Stats `json:"stats"`
	Version  interface{}         `json:"version"`
	Warnings []Warning           `json:"warnings"`
}

// QueryResponse_Stats defines model for QueryResponse.Stats.
type QueryResponse_Stats struct {
	ElapsedMs            int                    `json:"elapsed_ms"`
	Returned             *int                   `json:"returned,omitempty"`
	Scanned              *int                   `json:"scanned,omitempty"`
	AdditionalProperties map[string]interface{} `json:"-"`
}

// TraceTree defines model for TraceTree.
type TraceTree struct {
	// Spans The trace's spans in tree order (parent before children).
	Spans []map[string]interface{} `json:"spans"`

	// Trace opaque; see ../../model/v1alpha1/schema/trace.schema.json (generated types)
	Trace map[string]interface{} `json:"trace"`
}

// Warning Non-fatal signal using the llmobs.dq.* conventions (08-data-quality.md), e.g. a page that included truncated payloads or a partial result.
type Warning struct {
	Code    string                  `json:"code"`
	Detail  *map[string]interface{} `json:"detail,omitempty"`
	Message string                  `json:"message"`
}

// InvalidQuery defines model for InvalidQuery.
type InvalidQuery = Error

// NotFound defines model for NotFound.
type NotFound = Error

// OverLimit defines model for OverLimit.
type OverLimit = Error

// Unauthorized defines model for Unauthorized.
type Unauthorized = Error

// RunQueryJSONBody defines parameters for RunQuery.
type RunQueryJSONBody = map[string]interface{}

// WriteScoreJSONBody defines parameters for WriteScore.
type WriteScoreJSONBody struct {
	union json.RawMessage
}

// WriteScoreJSONBody0 defines parameters for WriteScore.
type WriteScoreJSONBody0 = map[string]interface{}

// WriteScoreJSONBody1 defines parameters for WriteScore.
type WriteScoreJSONBody1 = []map[string]interface{}

// EraseSpansParams defines parameters for EraseSpans.
type EraseSpansParams struct {
	UserId string    `form:"user_id" json:"user_id"`
	From   time.Time `form:"from" json:"from"`
	To     time.Time `form:"to" json:"to"`
}

// RunQueryJSONRequestBody defines body for RunQuery for application/json ContentType.
type RunQueryJSONRequestBody = RunQueryJSONBody

// WriteScoreJSONRequestBody defines body for WriteScore for application/json ContentType.
type WriteScoreJSONRequestBody WriteScoreJSONBody

// Getter for additional properties for QueryResponse_Stats. Returns the specified
// element and whether it was found
func (a QueryResponse_Stats) Get(fieldName string) (value interface{}, found bool) {
	if a.AdditionalProperties != nil {
		value, found = a.AdditionalProperties[fieldName]
	}
	return
}

// Setter for additional properties for QueryResponse_Stats
func (a *QueryResponse_Stats) Set(fieldName string, value interface{}) {
	if a.AdditionalProperties == nil {
		a.AdditionalProperties = make(map[string]interface{})
	}
	a.AdditionalProperties[fieldName] = value
}

// Override default JSON handling for QueryResponse_Stats to handle AdditionalProperties
func (a *QueryResponse_Stats) UnmarshalJSON(b []byte) error {
	object := make(map[string]json.RawMessage)
	err := json.Unmarshal(b, &object)
	if err != nil {
		return err
	}

	if raw, found := object["elapsed_ms"]; found {
		err = json.Unmarshal(raw, &a.ElapsedMs)
		if err != nil {
			return fmt.Errorf("error reading 'elapsed_ms': %w", err)
		}
		delete(object, "elapsed_ms")
	}

	if raw, found := object["returned"]; found {
		err = json.Unmarshal(raw, &a.Returned)
		if err != nil {
			return fmt.Errorf("error reading 'returned': %w", err)
		}
		delete(object, "returned")
	}

	if raw, found := object["scanned"]; found {
		err = json.Unmarshal(raw, &a.Scanned)
		if err != nil {
			return fmt.Errorf("error reading 'scanned': %w", err)
		}
		delete(object, "scanned")
	}

	if len(object) != 0 {
		a.AdditionalProperties = make(map[string]interface{})
		for fieldName, fieldBuf := range object {
			var fieldVal interface{}
			err := json.Unmarshal(fieldBuf, &fieldVal)
			if err != nil {
				return fmt.Errorf("error unmarshaling field %s: %w", fieldName, err)
			}
			a.AdditionalProperties[fieldName] = fieldVal
		}
	}
	return nil
}

// Override default JSON handling for QueryResponse_Stats to handle AdditionalProperties
func (a QueryResponse_Stats) MarshalJSON() ([]byte, error) {
	var err error
	object := make(map[string]json.RawMessage)

	object["elapsed_ms"], err = json.Marshal(a.ElapsedMs)
	if err != nil {
		return nil, fmt.Errorf("error marshaling 'elapsed_ms': %w", err)
	}

	if a.Returned != nil {
		object["returned"], err = json.Marshal(a.Returned)
		if err != nil {
			return nil, fmt.Errorf("error marshaling 'returned': %w", err)
		}
	}

	if a.Scanned != nil {
		object["scanned"], err = json.Marshal(a.Scanned)
		if err != nil {
			return nil, fmt.Errorf("error marshaling 'scanned': %w", err)
		}
	}

	for fieldName, field := range a.AdditionalProperties {
		object[fieldName], err = json.Marshal(field)
		if err != nil {
			return nil, fmt.Errorf("error marshaling '%s': %w", fieldName, err)
		}
	}
	return json.Marshal(object)
}

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Execute a query DSL document.
	// (POST /v1alpha1/query)
	RunQuery(w http.ResponseWriter, r *http.Request)
	// The score write path (requires the `scores:write` scope). Enforces LM-3/LM-8: measurement-only value model and (subject_type, subject_id) subjects, including plugin-namespaced subject types. Accepts a single score or a batch array; writes are synchronous so validation is observable (04-score.md §7). A batch is all-or-nothing.
	// (POST /v1alpha1/scores)
	WriteScore(w http.ResponseWriter, r *http.Request)
	// Fetch a single score by id.
	// (GET /v1alpha1/scores/{id})
	GetScore(w http.ResponseWriter, r *http.Request, id string)
	// GDPR erasure (requires the `delete` scope). Hard-deletes every span matching a mandatory, bounded filter — `user_id` plus a `timeRange` — and records an erasure audit row (actor, filter, count, time). Synchronous: the filter is bounded and the caller needs the proven count.
	// (DELETE /v1alpha1/spans)
	EraseSpans(w http.ResponseWriter, r *http.Request, params EraseSpansParams)
	// Fetch a single span by id.
	// (GET /v1alpha1/spans/{id})
	GetSpan(w http.ResponseWriter, r *http.Request, id string)
	// Fetch a single trace by id.
	// (GET /v1alpha1/traces/{id})
	GetTrace(w http.ResponseWriter, r *http.Request, id string)
	// Optimized full span-tree fetch (the reason for the tree-first sort key): the trace, its spans in tree order, and span events, deduplicated per LM-5.
	// (GET /v1alpha1/traces/{trace_id}/tree)
	GetTraceTree(w http.ResponseWriter, r *http.Request, traceId string)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// RunQuery operation middleware
func (siw *ServerInterfaceWrapper) RunQuery(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, ServiceTokenScopes, []string{})

	ctx = context.WithValue(ctx, UserAssertionScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.RunQuery(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// WriteScore operation middleware
func (siw *ServerInterfaceWrapper) WriteScore(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, ServiceTokenScopes, []string{})

	ctx = context.WithValue(ctx, UserAssertionScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.WriteScore(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetScore operation middleware
func (siw *ServerInterfaceWrapper) GetScore(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, ServiceTokenScopes, []string{})

	ctx = context.WithValue(ctx, UserAssertionScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetScore(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// EraseSpans operation middleware
func (siw *ServerInterfaceWrapper) EraseSpans(w http.ResponseWriter, r *http.Request) {

	var err error

	ctx := r.Context()

	ctx = context.WithValue(ctx, ServiceTokenScopes, []string{})

	ctx = context.WithValue(ctx, UserAssertionScopes, []string{})

	r = r.WithContext(ctx)

	// Parameter object where we will unmarshal all parameters from the context
	var params EraseSpansParams

	// ------------- Required query parameter "user_id" -------------

	if paramValue := r.URL.Query().Get("user_id"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "user_id"})
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "user_id", r.URL.Query(), &params.UserId)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "user_id", Err: err})
		return
	}

	// ------------- Required query parameter "from" -------------

	if paramValue := r.URL.Query().Get("from"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "from"})
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "from", r.URL.Query(), &params.From)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "from", Err: err})
		return
	}

	// ------------- Required query parameter "to" -------------

	if paramValue := r.URL.Query().Get("to"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "to"})
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "to", r.URL.Query(), &params.To)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "to", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.EraseSpans(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetSpan operation middleware
func (siw *ServerInterfaceWrapper) GetSpan(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, ServiceTokenScopes, []string{})

	ctx = context.WithValue(ctx, UserAssertionScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetSpan(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetTrace operation middleware
func (siw *ServerInterfaceWrapper) GetTrace(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, ServiceTokenScopes, []string{})

	ctx = context.WithValue(ctx, UserAssertionScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetTrace(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetTraceTree operation middleware
func (siw *ServerInterfaceWrapper) GetTraceTree(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "trace_id" -------------
	var traceId string

	err = runtime.BindStyledParameterWithOptions("simple", "trace_id", r.PathValue("trace_id"), &traceId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "trace_id", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, ServiceTokenScopes, []string{})

	ctx = context.WithValue(ctx, UserAssertionScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetTraceTree(w, r, traceId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of http.ServeMux.
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	ServeHTTP(w http.ResponseWriter, r *http.Request)
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc("POST "+options.BaseURL+"/v1alpha1/query", wrapper.RunQuery)
	m.HandleFunc("POST "+options.BaseURL+"/v1alpha1/scores", wrapper.WriteScore)
	m.HandleFunc("GET "+options.BaseURL+"/v1alpha1/scores/{id}", wrapper.GetScore)
	m.HandleFunc("DELETE "+options.BaseURL+"/v1alpha1/spans", wrapper.EraseSpans)
	m.HandleFunc("GET "+options.BaseURL+"/v1alpha1/spans/{id}", wrapper.GetSpan)
	m.HandleFunc("GET "+options.BaseURL+"/v1alpha1/traces/{id}", wrapper.GetTrace)
	m.HandleFunc("GET "+options.BaseURL+"/v1alpha1/traces/{trace_id}/tree", wrapper.GetTraceTree)

	return m
}